refactor(headless): extract adjudication helpers to _headless_adjudication.py
Move _resolve_skill_session_id/_parse_stdout/_build_api_retry_outcome/
_make_terminated_result/_has_out_of_cwd_file_change/_apply_post_session_adjudication/
_validate_declared_artifact from _headless_result.py (1030 → 846 lines)
to _headless_adjudication.py. _build_skill_result remains as the
headless orchestration authority. Re-export the moved symbols so
existing imports keep working. Narrow E25 line-cap to acknowledge
the residual 846-line _headless_result.py.

Refs: #4664
